tools/requirements.txt
openai
google-auth
google-auth-httplib2
google-api-python-client
requests
python-dotenv
pillow
